requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9 # For PostgreSQL connection
python-jose[cryptography]==3.3.0 # For JWT handling
passlib[bcrypt]==1.7.4 # For password hashing
python-dotenv==1.0.1 # For loading .env variables